+++ test/cases/js-import-css-modules-esm/src/main.js
@@ -1,0 +1,4 @@
+import tmpl from './partial.html';
+import styles from './style.css';
+
+document.getElementById('root').innerHTML = tmpl({ styles });